Ensure safe callback cleanup when dropping IndexReader (#10) (#14)
### Issue
After the IndexReader is dropped, the FileWatcher may continue running.
This can lead to lock files being generated in the directory, which may
cause Milvus to fail when attempting to delete the directory

### Solution
Ensure FileWatcher stops running when IndexReader is dropped, preventing
unnecessary lock file creation and avoiding deletion errors.

---------

Signed-off-by: sunby <sunbingyi1992@gmail.com>